Campfire and Soul Wisp Added
The campfire and soul wisps are using placeholder art for now. The campfire has a noticeable radius that when within, will sap the player for their souls at a rather quick rate. Soul Wisps will spawn to signify that souls are being sapped. They float towards and destroy themselves in the fire. A current bug exists where the Soul Wisps will continue to be sapped from the player even when the player exits the campfire's sapping range.